Split metadata module and make public
Move funtions that need to know about dataset to that module
and extend with a function that takes dataset rather than runid

Split Metadatable class and shapshot diff functions into different
submodules and make all public
