+++ qcodes/instrument/instrument_base.py
@@ -19,8 +19,7 @@ import numpy as np
 
 from qcodes.logger import get_instrument_logger
 from qcodes.parameters import Function, Parameter, ParameterBase
-from qcodes.utils import DelegateAttributes, full_class
-from qcodes.utils.metadata import Metadatable
+from qcodes.utils import DelegateAttributes, Metadatable, full_class
 
 if TYPE_CHECKING:
     from qcodes.instrument.channel import ChannelTuple, InstrumentModule
