+++ qcodes/loops.py
@@ -49,8 +49,7 @@ import numpy as np
 from qcodes.data.data_array import DataArray
 from qcodes.data.data_set import new_data
 from qcodes.station import Station
-from qcodes.utils import full_class
-from qcodes.utils.metadata import Metadatable
+from qcodes.utils import Metadatable, full_class
 
 from .actions import (
     BreakIf,
