+++ qcodes/parameters/function.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING, Any, Callable, List, Optional, Sequence, Union
 
-from qcodes.utils.metadata import Metadatable
+from qcodes.utils import Metadatable
 from qcodes.validators import Validator, validate_all
 
 from .command import Command
