+++ qcodes/tests/common.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Any, Callable, Dict, Hashable, Optional, Tuple
 
 import qcodes
 from qcodes.configuration import Config, DotDict
-from qcodes.utils.metadata import Metadatable
+from qcodes.utils import Metadatable
 
 if TYPE_CHECKING:
     from pytest import ExceptionInfo
