+++ qcodes/utils/__init__.py
@@ -17,16 +17,20 @@ from .installation_info import (
     is_qcodes_installed_editably,
 )
 from .json_utils import NumpyJSONEncoder
+from .metadatable import Metadatable
 from .numpy_utils import list_of_data_to_maybe_ragged_nd_array
 from .partial_utils import partial_with_docstring
 from .path_helpers import get_qcodes_path, get_qcodes_user_path
+from .snapshot_helpers import ParameterDiff, diff_param_values, extract_param_values
 from .spyder_utils import add_to_spyder_UMR_excludelist
 from .threading_utils import RespondingThread, thread_map
 
 __all__ = [
     "DelayedKeyboardInterrupt",
     "DelegateAttributes",
+    "Metadatable",
     "NumpyJSONEncoder",
+    "ParameterDiff",
     "QCoDeSDeprecationWarning",
     "RespondingThread",
     "attribute_set_to",
@@ -34,6 +38,7 @@ __all__ = [
     "convert_legacy_version_to_supported_version",
     "deep_update",
     "deprecate",
+    "diff_param_values",
     "full_class",
     "get_all_installed_package_versions",
     "get_qcodes_path",
